refactor: split Header.tsx (385→168 LOC) into StatusIndicators and QACoverageBadge
Extract PhaseBadge, StatusDot, ConnectionIndicator, ElapsedTimer to
StatusIndicators.tsx (~98 LOC) and QACoverageBadge with its types/helpers
to QACoverageBadge.tsx (~142 LOC). Header.tsx now imports from both and
re-exports all symbols for backward compatibility. All 257 tests pass.
